Split evidence source registry catalogs
